fix incorrect prop, rename flags reducer to status
