+++ packages/suite/src/middlewares/wallet/walletMiddleware.ts
@@ -6,40 +6,37 @@ import * as walletActions from '@wallet-actions/walletActions';
 import * as discoveryActions from '@wallet-actions/discoveryActions';
 import { SUITE } from '@suite/actions/suite/constants';
 import { WALLET } from '@wallet-actions/constants';
-import { getRouteFromPath } from '@suite/utils/suite/router';
 import { AppState, Action, Dispatch } from '@suite-types';
 
+// Flow: LOCATION.CHANGE -> WALLET.INIT -> load storage -> WALLET.INIT_SUCCESS
 const walletMiddleware = (api: MiddlewareAPI<Dispatch, AppState>) => (next: Dispatch) => (
     action: Action,
 ): Action => {
     const prevState = api.getState();
     // pass action
     next(action);
-    // run only in wallet app
-    if (prevState.router.app !== "wallet")
-        return action;
-    
+
+    // runs only in wallet app
+    if (prevState.router.app !== 'wallet') return action;
+
     switch (action.type) {
         case LOCATION_CHANGE: {
+            // update selected account if needed
             api.dispatch(selectedAccountActions.observe(prevState, action));
-            const route = getRouteFromPath(action.url);
-            console.log(route);
-            if (route && route.pattern.startsWith('/wallet')) {
-                api.dispatch(walletActions.init());
-            }
+
+            // dispatch wallet init, then load storage
+            api.dispatch(walletActions.init());
             break;
         }
 
         case WALLET.INIT:
             api.dispatch(loadStorage());
             break;
-            
-            case SUITE.SELECT_DEVICE:
-            case SUITE.UPDATE_SELECTED_DEVICE:
-                if (api.getState().suite.device) {
-                    api.dispatch(discoveryActions.start())
-                };
-                api.dispatch(selectedAccountActions.observe(prevState, action));
+
+        case SUITE.SELECT_DEVICE:
+        case SUITE.UPDATE_SELECTED_DEVICE:
+            api.dispatch(discoveryActions.init());
+            api.dispatch(selectedAccountActions.observe(prevState, action));
             break;
         default:
             break;
