+++ packages/suite/src/reducers/wallet/index.ts
@@ -9,7 +9,7 @@ import discoveryReducer from './discoveryReducer';
 import accountsReducer from './accountsReducer';
 import selectedAccountReducer from './selectedAccountReducer';
 import receiveReducer from './receiveReducer';
-import flagsReducer from './flagsReducer';
+import statusReducer from './statusReducer';
 
 const WalletReducers = combineReducers({
     signVerify: signVerifyReducer,
@@ -22,7 +22,7 @@ const WalletReducers = combineReducers({
     accounts: accountsReducer,
     selectedAccount: selectedAccountReducer,
     receive: receiveReducer,
-    flagsReducer: flagsReducer,
+    status: statusReducer,
 });
 
 export default WalletReducers;
