+++ packages/suite/src/views/wallet/account/summary/Container.ts
@@ -8,6 +8,7 @@ import Summary from './index';
 
 const mapStateToProps = (state: AppState) => ({
     wallet: state.wallet,
+    suite: state.suite,
     fiat: state.wallet.fiat,
     router: state.router,
 });
